+++ pkg/login/access_test.go
@@ -1,4 +1,4 @@
-package access
+package login
 
 import (
 	"github.com/stretchr/testify/assert"
