+++ pkg/login/session_test.go
@@ -1,4 +1,4 @@
-package access
+package login
 
 import (
 	"context"
