+++ pkg/login/sessions.go
@@ -1,4 +1,4 @@
-package access
+package login
 
 import (
 	"bytes"
